+++ modules/__init__.py
@@ -160,13 +160,13 @@ sitesRankCacheTimeout = 300
 sitesRankCacheStale   = 3600
 
 workflowStates = [ 'draft','submitted','approved','running',
-                   'paused','checking','finished','deleted']
+                   'paused','checking','finishing','finished','deleted']
 
 jobStatesTerminal = [ 'finished', 'notused', 'aborted', 'stalled', 
                       'jobscript_error', 'outputting_failed', 'none_processed' ]
 
-jobStatesAll = [ 'submitted', 'started', 'processing', 'outputting' ] \
-               + jobStatesTerminal
+jobStatesAll = [ 'submitted', 'started', 'processing', 'outputting', 
+                 'finishing'] + jobStatesTerminal
 
 jobStallSeconds = 3660
 maxFilesPerJob  = 20